requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
